+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.47.1 - TBD
+## 0.48.0 - 2026-02-18
 
 ### Enhancements
 - Added `SlowReaderBehavior` enum and `LiveBuilder::SetSlowReaderBehavior()` to configure
@@ -15,7 +15,7 @@
 
 ### Enhancements
 - Added Zstd compression support to live clients which can be enabled with
-`LiveBuilder::SetCompression()`. It's disabled by default
+  `LiveBuilder::SetCompression()`. It's disabled by default
 - Added `Compression()` getter to `LiveBlocking` and `LiveThreaded`
 - Upgraded default `httplib` version to 0.30.1